.windsurf/autoclicker/requirements.txt
Pillow>=10.0.0
opencv-python>=4.8.0
PyAutoGUI>=0.9.54